fix warnings change way folders are created

+++ src/core/soilmoisture.cpp
@@ -416,7 +416,7 @@ FrostComponent::getMeanBulkDensity()
 {
   auto vs_number_of_layers = soilColumn.vs_NumberOfLayers();
   double bulk_density_accu = 0.0;
-  for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
+  for (int i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
     bulk_density_accu += soilColumn[i_Layer].vs_SoilBulkDensity();
   }
   return (bulk_density_accu / double(vs_number_of_layers) / 1000.0); // [Mg m-3]
@@ -431,7 +431,7 @@ FrostComponent::getMeanFieldCapacity()
 {
   auto vs_number_of_layers = soilColumn.vs_NumberOfLayers();
   double mean_field_capacity_accu = 0.0;
-  for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
+  for (int i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
     mean_field_capacity_accu += soilColumn[i_Layer].vs_FieldCapacity();
   }
   return (mean_field_capacity_accu / double(vs_number_of_layers));
@@ -610,9 +610,9 @@ FrostComponent::updateLambdaRedux()
 {
   auto vs_number_of_layers = soilColumn.vs_NumberOfLayers();
 
-  for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
+  for (int i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
 
-    if (i_Layer < (size_t(std::floor((vm_FrostDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
+    if (i_Layer < (std::floor((vm_FrostDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5))) {
 
       // soil layer is frozen
       soilColumn[i_Layer].vs_SoilFrozen = true;
@@ -623,7 +623,7 @@ FrostComponent::updateLambdaRedux()
       }
     }
 
-    if (i_Layer < (size_t(std::floor((vm_ThawDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
+    if (i_Layer < (std::floor((vm_ThawDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5))) {
       // soil layer is thawing
 
       if (vm_ThawDepth < (double(i_Layer + 1) * soilColumn[i_Layer].vs_LayerThickness) && (vm_ThawDepth < vm_FrostDepth)) {
@@ -652,7 +652,7 @@ FrostComponent::updateLambdaRedux()
       vm_FrostDays = 0;
 
       vm_HydraulicConductivityRedux = pm_HydraulicConductivityRedux;
-      for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++)
+      for (int i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++)
       {
         soilColumn[i_Layer].vs_SoilFrozen = false;
         vm_LambdaRedux[i_Layer] = 1.0;
